Add files via upload
Emissivities of 50/50 VolFrac Dirty Ice matrix with AstroSil inclusion.
Graph is sampling of emissivities of grains of radii expressible as integer powers of 10
Unfortunately the legend is placed poorly; even more unfortunate is that running MIE takes about 1 hour 3 minutes so running it again just to generate a new plot is a silly thing. When the reader for the emiss table is ready it will be used to read in said emissivities and make a new plot.