Add files via upload
Hey Jon, these two files are what we talked about yesterday. temptable.50-50 is pickled file that contains two elements. A list of temperatures and a dictionary holding the radius and grainsize associated with those temperature. Key = temp, values = radius, grainsize.

The second file pickleload is a short snip of code that will load these parameters into your program, plus a short example of how the dictionary can be used. I tried to comment deeply in that file but if the comments don't help just ask me. 

I did not include emissivity in the dictionary because emissivity ends up being a factor of wavelength as well as grainsize, plus it's already factored into the flux interpolation that derives the temperature.